Store only one differntial in the MatrixComplex and rewrite resize.
Thus we may re-use the storage of the (n-1)th differential
for the nth differential.

NOTE: Several methods are renamed or removed. One might
want to write methods to compute homology later.

In order to avoid runtime wasting reallocations, we reuse
the storage of the (n-1)th differential for the nth differential
by using resize-functions of vector and dynamic_bitset.