Radial Case: Change definition of Tuple::operator bool.
Up to now, the boolean operator of a tuple returns true
iff the tuple does not contain the symbols 0 and p.

This cannot be used for radial cells. Thus there is now a new
way to construct Tuples, namely a constructor that initializes
a Tuple to contain no transposition at all. The operator bool
now checks whether the Tuple contains a transposition or not,
and we mark degenerate Tuples using the new constructor.